끝까지해내미_김민지_완료/requirements.txt
# AllerAlert Requirements for PythonAnywhere
# Install with: pip install -r requirements.txt

# Flask Web Framework
Flask==2.3.3
flask-cors==4.0.0

# Security and Authentication
Werkzeug==2.3.7

# HTTP Requests (for Gemini API)
requests==2.31.0

# Image Processing
Pillow==10.0.0

# Environment Variables
python-dotenv==1.0.0

# Optional: Production Server (not needed on PythonAnywhere)
# gunicorn==21.2.0